# openapi.yaml
openapi: 3.1.0
info:
  title: ForemanOS API
  version: 1.0.0
  summary: AI-powered construction project management API
  description: |
    AI-powered construction project management platform API.

    **Authentication:** Most endpoints require JWT authentication via NextAuth.

    **Rate Limits:**
    - Chat API: 20 requests/minute
    - Upload API: 10 requests/minute
    - Auth API: 5 requests/5 minutes
    - General API: 60 requests/minute

    **Subscription Tiers:** Free, Starter, Pro, Team, Business, Enterprise
  contact:
    name: ForemanOS Support
    url: https://foremanos.site
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://foremanos.site
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Request password reset
      description: |
        Send a password reset email to the user.

        **Security:**
        - Always returns success to prevent email enumeration
        - Rate limited to 5 requests per 5 minutes
        - 24-hour token expiration
        - Invalidates previous reset tokens
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  description: Email address of the account
                  example: "user@example.com"
      responses:
        '200':
          description: Success response (always returned to prevent enumeration)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "If the email exists, a password reset link has been sent."
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/reset-password:
    post:
      tags:
        - Auth
      summary: Reset password with token
      description: |
        Reset user password using a token from the forgot-password email.

        **Password Requirements:**
        - Minimum 8 characters
        - At least one uppercase letter
        - At least one lowercase letter
        - At least one number
        - At least one special character
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - password
              properties:
                token:
                  type: string
                  description: Reset token from email
                  example: "a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6"
                password:
                  type: string
                  format: password
                  description: New password meeting requirements
                  example: "NewP@ssw0rd123"
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password reset successful. You can now login with your new password."
        '400':
          description: Invalid request or token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid or expired reset token"
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/verify-email:
    get:
      tags:
        - Auth
      summary: Verify email address
      description: |
        Verify a user's email address using a token from the verification email.

        **Effects:**
        - Marks email as verified
        - Approves the account
        - Activates client role
      operationId: verifyEmail
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Verification token from email
          example: "v1e2r3i4f5y6t7o8k9e0n"
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Email verified successfully! You can now login."
                  User:
                    type: object
                    properties:
                      email:
                        type: string
                      username:
                        type: string
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid or expired verification token"
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: NextAuth JWT token from session
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Message or image is required"
...
    RateLimitExceeded:
      description: Too many requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds until retry is allowed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Too many requests. Please slow down and try again in a moment."
              retryAfter:
                type: integer
                description: Seconds until user can retry
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Internal server error"
              details:
                type: string
                description: Additional error details (dev mode only)
...